requirements.txt
Addons-Tkinter==0.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altgraph==0.17.5
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
arithmetic==0.6.2
asgiref==3.11.0
asyncpg==0.30.0
attrs==25.4.0
auto-py-to-exe==2.48.1
azure-cognitiveservices-speech==1.46.0
azure-core==1.36.0
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
distro==1.9.0
dj-config-url==0.1.1
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
dlib==20.0.0
Eel==0.18.2
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.120.1
fire==0.7.1
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llama_stack==0.3.1
llama_stack_client==0.3.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
media==0.1.6
multidict==6.7.0
numpy==2.3.4
openai==2.7.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==26.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
patsy==1.0.2
pefile==2024.8.26
pillow==12.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyaml==25.7.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyJWT==2.10.1
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
PyQt6_sip==13.10.3
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qtgui==0.0.1
qtwidgets==1.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
sc-pylibs==0.0.0.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
shiboken6==6.10.1
six==1.17.0
sklearn-linear-model-modification==0.0.11
sniffio==1.3.1
spglib==2.7.0
SQLAlchemy==2.0.44
sqlparse==0.5.5
starlette==0.49.1
statsmodels==0.14.5
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
utils==1.0.2
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.11.0
yarl==1.22.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
